Fix up integer constants to match magic rule: a normal decimal integer that is too long to fit in an 'int' will be promoted to 'long' even if it were to fit in 'unsigned int'. But hex and octal numbers stay as 'unsigned int' if they fit.
In other words, on a 32-bit architecture with a 32-bit long,
the constant '2147483648' gets implicitly converted first to
'long' because it doesn't fit in an 'int', and then to 'unsigned
long' because it doesn't fit in a 'long' either.

But the hex constant 0x80000000 (which has the same value),
will get converted to just an 'unsigned int'. Whee.