Show the expression parse tree as a sick kind of assembly language, which bears almost no resemblance to any _sane_ assembly language, but which gives a much better idea of what a back-end might actually do.
It also (since it has to linearize function calls and get argument sizes right
and so on) ends up being a pretty good way to see obvious breakage in the parse
tree. If you can just ignore the strangeness of the "assembly language".

My favourite "instructions":

	[vP]	cmov vT,vA,vB

here vP is a virtual register that acts as a "predicate" for the cmov, and
it will move the value of vA into vT if vP is non-zero, otherwise vB.  It
matches the C expression vT = vP ? vA : vB.

It also has the "notzero" instruction:

	notzero vT,vA

which sets vT to 0/1 depending on whether vA is nonzero/zero. It matches
the C expression vT = !vA.

Think of it as a three-register machine designed for REALLY stupid C compilers ;)