[PATCH] escape character extension
I added cases for the escape characters that show up in the 2.5.x kernel
sources.

One particular case (\<newline>), is a little more interesting.

This is different from the other cases because the \<newline> character
should never appear in the parsed string.

As a first fix, I've made escapechar() call itself recursively.  One
flaw with this fix is that escapechar() will mark the first character on
the second line of the input string as "escaped":

	/* Mark it as escaped */
	value |= 0x100;

but none of the callers to escapechar() ever use or preserve this
information, so it might as well be dropped altogether or at least
ignored for now.