[PATCH] another minor sparse change adding opnames.
Add more operations to the name list.  Without these additions,
the "and" operator is displayed as a null pointer.

In the original code, the "&" operator was within the name[] array,
but didnt' have a string assigned to it.  So printf() displayed it as
the string "null".  While I was at it, I added other operation names to
the table.