This makes function inlining largely work correctly. In order to get all the symbol replacement stuff right, we just build a per-function list of all the symbols associated with a function (except for arguments, which are special anyway due to the site-specific initializers).
When we inline, we then create the replacement symbols up front,
which makes it trivial to replace them as needed. This means that
we get all the special symbols (return targets etc) automatically
right.

Also fix a few problems with builtin symbols that were uncovered
by the inliner when running over the kernel sources.