[PATCH] Make sparse understand complex initializers inside expressions
This patch makes a cast expression followed by an initializer list into
a postfix expression that can be dereferenced as a structure or an array.
There are approximately 7 instances of these expressions in the Linux kernel,
that give warnings about "expected lvalue for member dereference".

The approach involved introducing a new "stack-based temporary" symbol
of the same type as the cast expression, and using this as the target
of the initialization expression.  The subsequent array or structure
member dereferences are made to that temporary symbol.

show-parse.c need modification to display a symbol expression with
an initializer list that is NOT in a symbol declaration list.

An example of this form with structure member dereference is:

    typedef struct { long t1; long t2; long t3; } longstruct_t;

    long test;

    int
    main(void)
    {
	int a, b, c;

	test = (longstruct_t){a, b, c}.t3;
	return 0;
    }

An example of this form with array member dereference is:

    typedef int iarray[2];
    int pgp;

    main(void)
    {
	int index;
	int a, b;

	pgp = (iarray){a,b}[index];
    }