[be] Don't emit move based on value returned from a function's toplevel STMT_COMPOUND.
If a function returns a value, then the C code must have explicit
'return' statements.  Since we already emit a move when handling
return statements, that eliminates the need to ever emit a move for
the non-'return' case.

Of course, incorrect C code can still omit the return statement, in
which case you're left with a random %eax return value.