[PATCH] PATCH: remove dead while loop
This improves the dead code eliminate for iterators that begin in an
unreachable section (eg a "while" statement inside a "if (0)").

The idea is that have a decoy entrypoint for those while loops.  On copy
the basic block back only if there is C level label was found inside the
block.

It is tempting that left the dead code elimination and constant
propagation to the some back end pass.