Remove bogus overly permissive type-comparisons.
We used to do bad things about pointers to arrays, and
as a result sparse had a bogus "pointers to arrays
degenerate to pointers to the array entry" special case
in the type equality checking.

However, it's not true that a pointer to an array is
type-equivalent with the pointer to an entry. Now that
we do degenerate arrays correctly, we should remove the
bogus special case.