This should largely fix token pasting.
It's not perfect now either, but it makes a hell of a lot more
sense now, and gets all the kernel cases I've tested right. Which
is actually saying quite a lot, since the kernel uses things that
are arguably not standard (the token fragments it pastes are not
valid tokens on their own, and I think that may be borderline in
ANSI - who knows).

It still doesn't do the _really_ strange cases, ie trying to
paste 

	a ## 0.e10

and expecting that to become "a0" "." "e10" is just not going
to happen. I think that's undefined behaviour anyway.