[PATCH] lazy-copy macro expansion in pre-processing
This is almost a total rewrite of the macro expansion.  It now does
lazy-copy or the arguments, with the realization that if an argument
only shows up once in the macro expansion (quite common), we don't need
to allocate a new copy of the argument token stream, we can just insert
it as-is.

I hope it's reasonably clean and readable; review would be obviously welcome.