Macroize lr_binop_expression() helper function.
This function was the top performance offender in sparse.

We really want it inlined, since all it really does is
to do an indirect call or two, and if inlined that call
turns into a direct call.

Making it a macro allows us to pass in the operation
comparison as code rather than as a list of integers,
which again allows the compiler to do a much better
job.