Warn about plain integer conversion to NULL pointer.
It tends to be a sign of type confusion. Yes, it's valid
C. And yes, some misguided people think it's the only true
C. But they are wrong.

NULL is "((void *)0)", and dammit, we should get a warning
if we use an integer as a pointer, or NULL as an integer.