[PATCH] teach sparse about __alignof__
This teaches sparse what __alignof__ really means, instead of just using
the same code as "__sizeof__"

It gets rid of the warnings in ebtables that does:
	struct ebt_entries {
...
	char data[0] __attribute__ ((aligned (__alignof__(struct ebt_replace)));
	};
Which caused warning because sparse was evaluating __alignof__ as the same as sizeof,
and sizeof was 57 (ie non-power of 2).

This is just based on the existing code and a peek at the data
structures in expression and symbol.