[PATCH] evaluate_conditional() prepared for FP
preparation for FP - we will be replacing conversions from FP to boolean
with explicit foo != 0.0 and !foo with foo == 0.0; that way a bunch of places
won't have to think of FP at all and uses of FP in boolean contexts are
rare enough to warrant that.  So we switch evaluate_conditional() to getting
a struct expression **, which will allow such rewriting.