Remove premature tree optimization on member accesses.
When the member offset was zero, we didn't create a subnode
to calculate the member address, since we already had it: it
is the same as the structure address.

However, while that is a valid optimization from a value
standpoint, it screws up the type system, since it overwrites
the structure pointer type with the member type, and thus
loses that step in the chain. This bites us later, when we
try to figure out what the tree means.

Just remove the optimization.