Parse K&R function declarations.
HOWEVER! We don't actually go and apply the types to the
arguments yet, so this is purely a "parse only" fix. To
actually make K&R functions work, we'd need to match up
the re-declared arguments against the argument list.

All the information is there now, though. Somebody more
interestedin K&R could fix it up.