[PATCH] uninlining inline functions
When we take the address of an inline function or otherwise refusing to
inline it, we need to output the now non-inline function properly.

What we do is
	a) keeping body and symbol list of inlined function in new fields
	b) when expanding inlined call use these fields
	c) when evaluating the function itself (which happens if sparse
decides that it can't be [always] inlined) uninline the sucker.  I.e.
create ->stmt and ->symbol_list by copying the ->inline_stmt and
->inline_symbol_list same as we would do while expanding a call.

That guarantees that we won't run into trouble with inlined calls coming
afterwards - evaluation doesn't mangle ->inline_stmt anymore.