[PATCH] __attribute__((bitwise))
Handling of __attribute__((bitwise)) in a way that should be easy to extend
afterwards.   Example of use:

typedef __u32 __attribute__((bitwise)) __le32;

That will create a new 32bit type that will be assignment-incompatible with
anything else.  The set of allowed operations is restricted to bitwise ones,
the only allowed constant is 0 right now.  Forced casts are allowed, so is
cast from type to itself and cast to void.  Any other cast will give a warning.

Checks are triggered by -Wbitwise in command line; if it's not there,
attribute will be silently ignored.