Complain about type differences in signedness.
We've got a fair number of these in the kernel, because
gcc historically hasn't checked it right, so lots of small
problems have crept in. But gcc-4 will check, and we really
should clean it up.

Make the warning string a bit more readable for the sign-only
difference case. NOTE! We're very anal about these things because
we check for _explicit_ signedness differences too. That might
be overkill, of course.