Totally re-do how we build up the initializer tree: make the positional markers be hierarchical rather than a flat list.
This makes the data structure a bit more complex, but it simplifies
some of the code, and makes it possible to evaluate complex initializers
without going insane.

In particular, we how support

	struct xxxx var = {
		.a.b[10] = 1;
	};

which we couldn't handle at all before (it had to
be written as

	struct xxxx var = {
		.a = {
			.b = { [10] = 1; }
		}
	}

or similar.

The new code changes all array indexes and structure members
to EXPR_POS expressions offset from the "outer" scope (either
start of the symbol, or an outer EXPR_POS).