[PATCH] WinGW/CygWin compatibility support
First cut at CygWin/MinGW compat support.

Currently, for mingw, I am using windows API to identify the files
uniquely, which is somewhat heavy because I have to open both the files
simultaneously to compare the indices, will replace it to a lighter
version later if I find out how to do that.