[PATCH] enum handling
a) we allow enum have an equivalent type other than int (it is kept in
   ->ctype.base_type for SYM_ENUM types).  Code adjusted.

b) enum declaration parsing tries to determine the equivalent type; if
   such type does not exist we set it to &bad_enum_ctype and
   examine_symbol_type() generates a warning when it meets such beast.
   The rules for equivalent type follow:

	1) if all members have the same bitwise type, it is the
	   equivalent type of enum.
	2) if all members have integer types and there is a single
	   integer type capable of representing all their values, we pick
	   the first such type (in the same ordering that is used for e.g.
	   integer constants). 
	3) anything else => bad_enum_ctype.

c) enum is compatible with its equivalent type. 

d) it is an error to have implied initializer for a member following
   that of bitwise type.

There are still issues with float in enums, but since gcc would choke on
those unconditionally, we can be sure that there's no instances in any
code we could deal with (unless we start playing with Plan 9 codebase,
and that would require more work in other places - float enum is not the
only C extension in there).  Separate patch...