Add __sizeof_ptr__ that looks at a pointer expression and returns the size of the underlying object.
This is different from "sizeof(*expr)" for arrays, where
the array would degenerate to a pointer to one member, and
thus "sizeof(*expr)" gives the size of one entry in the
array.

Why do this? It's useful for things like

   #define memset(a,b,c) ({ \
        (void) __builtin_warning(__sizeof_ptr__(a) > 1, __sizeof_ptr__(a) != (c), "check memset size"); \
        memset(a, b, c); })

where we really want to check the size of the object we're
doing the "memset()" on, but the regular sizeof() just doesn't
cut it.