Clean up declarator sym->ident handling.
Since bind_symbol() now sets it, we don't need to set it by
hand any more.

Also, we really don't need to pass a pointer to a symbol
pointer in the declarator parsing functions, since they
never change it, and depend on it being set anyway. Might
as well just pass in the direct pointer to the symbol
instead.