Replace context/contextmask (never used) with in_context/out_context.
It's a way of telling the checker what the input context count is
expected to be, and what the output context should be.

For example, a function that is expected to be called with a spinlock
held, and releases that lock, could have an in_context of 1, and an
out_context of 0.