Do silly phi-node "expansion" as the last phase of the linearized tree.
This just copies all the phi node information into the
source of the phi node, so that the linearized thing looks
a bit more like "real" assembly language. It's likely
bogus in many ways, but I want to make a pretty-printer
for the linearized output, and this will make it much
much easier to do.