Introduce "value pseudos" and implicit deathnotes.
A "value pseudo" is a pseudo that just has a constant value
at any particular time. Instead of allocating a real pseudo
for it, and associating the pseudo with an expression, the
pseudo itself is directly associated with the value.

Also remove a few explicit death-notes: branches and calls
always kill their arguments, so make that implicit rather
than waste time addign explicit notes.