Pack basic blocks only after phi node operations.
The basic block packing may remove blocks that are unreachable,
but that are the source of a phi-node. That confuses the phi
node simplification code, and the simple solution is to just
delay the packing.

That's doubly true, since branch following in phi node simplification
may actually result in unreachable basic blocks that should be
packed away.