Remove EXPR_BITFIELD entirely.
I used to think I needed it. That's no longer the case: we just
follow the "bit_offset" in the type information.

There may be cases where we inadvertently cast the information
away, and those places will break now, but that's a bug really,
not an excuse for EXPR_BITFIELD.