Make the pseudo usage list be a list of pointers to the pseudos, not the instructions that use them.
That makes replacing a pseudo trivial: just walk the
list, and overwrite the thing we have a pointer to.

We also take advantage of the fact that symbol pseudos
only exist in the "address" part of memory load, so
if we know the pointer to the pseudo, we can calculate
the pointer to the instruction.

This makes the symbol pseudo replacement a lot more
effective.