Honor pre-processor conditionals even when they don't match within a stream.
We still warn about them, but we act as if it was ok.

The "vik2" IOCCC-2004 entry does some nasty preprocessor stuff on
purpose. We parse it all right, but because we refused to honor #if
statements that are terminated in another stream, we didn't get the
output Daniel Vik intended.