Make the pre-processor free the tokens that never make it into the token stream.
Most preprocessor lines are unused after handling, with the
notable exception of a #define. And regular tokens that get
skipped due to being in a region that is inside a false #if
case are also obvious fodder for this.

This shrinks the memory usage by a factor of five on some
particularly nasty code in the 2004 IOCCC.