Replace OP_LOAD/OP_STORE with OP_LNOP/OP_SNOP when making them irrelevant.
This shows an interesting (?) bug with uninlining, where we seem to
go through the same symbol twice, resulting in complaints about symbols
used in strange ways (because we see an instruction that we have already
NOP'ped out, using a symbol).

I'm not convinced this is a bug yet - it might just be that we really
_can_ reach the same symbol twice. But I _think_ it means our uninliner
may have problems.