Clean up pseudo and bb usage handling.
We really don't care about all bb usage, only about the phi-nodes
that use bb's. All the other usage is pretty obvious from the
flow graph, no need to add them to the list (and if we _only_ add
phi-nodes, then it becomes easier to check whether a bb needs to
be saved due to phi-node issues).

The pseudo usage has nothing to do with an instruction, so don't
mix in the instruction into it.