Remove "struct phi", replace with instruction that generates a pseudo.
The instruction contains everything "struct phi" had, namely
source bb and pseudo. And generating a new pseudo means that
we not only have the pointer to the instruction that defines
the phi-node (in pseudo->def), we also end up with automatic
phi-node usage tracking (pseudo->users).

The changes may look large, but almost all of this is just
direct search-and-replace of the old phi-node usage.