Move instruction simplification to new file "simplify.c".
Also, allow marking a pseudo unused by setting it to VOID,
which just disables further renaming of that use. This is
needed to make phi-source instructions (that can be shared
among multiple phi-nodes thanks to CSE) be collectable.

We used to just mark the phi source dead, but that was
wrong, since _another_ phi node might be using it.