Change linearizer debug output syntax for call instructions.
It now ends up being

	CALL fn, arg, argc, arg
	pseudo <- %retval

for a function that returns a return value.

This is slightly closer to what real assembly would be like,
but more importantly, it removes the only valid use of VOID
pseudos from the printout, so now you can see when something
goes wrong y just grepping for VOID.

This happens in the kernel a lot. Apparently mainly because
of lack of support for inline assembly.