Fix subtle problem with fn array arguments
We used to (long ago) consider a zero bit_size to be a sign
of the symbol type needing to be examined. However, when I
changed that to use an explicit "examined" flag, I didn't
realize that the function argument degeneration had depended
on that.

So make the function argument degeneration explicitly ask for
re-examination after changing it to a pointer. Otherwise it
ends up being a "zombie pointer" of size 0, which makes for
lots of confusion.