Leave symbol pseudo usage intact when doing phi-node conversion.
There's actually a reason we leave a bogus usage list for symbol
pseudo's: they are special, and they traverse the list _while_
doing cleanup, and changing the list while they are traversing
it is a bad idea.

Also, symbol pseudos have special usage rules, and know "where"
the usage is in an instruction, so they can (and do) check when
their usage has changed.