Associate pseudos with the symbol name whose value they got.
This is purely for debugging. It's only used to show what symbol
a pseudo might be (and I stress "might", since CSE can and does
screw it up) associated with.

Also print out the def list for a basic block when verbose.

It all makes it a bit easier to guess what's up.