Simplify trivial casts (and handle pointers specially).
This does trivial simplification of casting to the same
typesize. HOWEVER. We split casts up into whether they
cast to a dereferencable pointer type or not, and we don't
simplify pointer casts. This should mean that if we
ever want to do type-based alias analysis, we can still
avoid casted accesses.

(If we do type-based alias analysis, we'll also need to make
a union access do a cast. Which we probably should do anyway).