Add pseudo death-note tracking.
Now that we have full pseudo usage lists, we can also add
deathnotes to pseudos in the instruction stream.

NOTE! We add the deathnote to _before_ the last instruction
that uses that pseudo. That looks a bit strange, but it's
actually what you want: when we traverse the instuctions, we
want to know that the inputs are dead.