Remove stat-based file identity tests.
Replace it with a simple pathname comparison instead.

The pathname check is not only portable (no need for any
compatibility helper functions), but we can do it much
earlier, and thus make the check much cheaper by avoiding
three extra system calls when it triggers (open/fstat/close).

And the pathname test seems to match all the cases anyway.