Add storage usage verification phase.
Since we don't actually verify that storage doesn't overlap
when we do code generation, out local heuristics sometimes
do the wrong thing and allocate the same register to two
pseudos. This should catch it.

We do pretty well most of the time, so it's not obvious from
just reading the output. Largely because we try to do parents first,
which means that inputs are usually stable.