Make "sparse()" handle multiple input files on the command line
Pretty bogus right now, but at least it does more of the argument
parsing, and the infrastructure is there to handle multiple files.

It even works, sometimes.  Mostly not, though - since it will actually
use global scope for global types, and complain about redefinitions of
the same struct when it sees ti in two different compilation units.