Don't overwrite built-in macros when redefining them.
This means that we may have both an "invisible" macro _and_ a visible
one in scope, so we need to be a bit more careful about lookup.

We always look up the topmost one (either visible or invisible), and if
it's invisible, that means it is undefined (a builtin macro was first
re-defined, and then that redefinition was undefined).