[PATCH] Fix address space ordering problem
We used to get "__user void *" wrong - it ended up with address space 0
instead of address space 1.

What happens is actually pretty simple - we get address_space(1) handled
in declaration_specifiers(), which sets ctype->as to 1.  Then we see
"void" and eventually get to

                        ctype->base_type = type;
                }

                check_modifiers(&token->pos, s, ctype->modifiers);
                apply_ctype(token->pos, &thistype, ctype);

with thistype coming from lookup for "void".  And that, of course, has
zero ->as.  Now apply_ctype merrily buggers ctype->as and we have 0...

So AFAICS proper fix for sparse should be to check thistype->as to see
if it really has any intention to change ->as.

Signed-off-by: Linus Torvalds <torvalds@osdl.org>