Do stupid and crappy CSE on casts.
It's better than nothing, but I need to re-do cast linearization some
day.  Right now we squirrell away the whole original type, but the fact
is, we only need to save away a few bits of "type of cast".

We need to know if it's sign-extending, and whether the source or
destination are floating point values.  And the source size, of course.
But we don't really need to know the detailed original type.

Oh, well. This makes at least trivial things look much better.