Make sure we keep enum values in a sufficiently large type for parsing
We do enum type handling in two passes: one while parsing the values,
and then afterwards we determine the final type that depends on the
range of the results.

Make sure that the intermediate stages keep the intermediate types big
enough to cover the full range.

(The final type-casting is also buggy, but that's a separate issue)

Signed-off-by: Linus Torvalds <torvalds@osdl.org>