Cast all enum values to the final type
The type of an enum is determined by the type of all of its entries,
which means that while we may parse it in one type, it might end up with
another type in the end.  We used to just switch the types around, but
that didn't properly upgrade the actual values to the new type.

The trivial fix is to just keep a list of entries around, and then go
back and cast the values at the end.

Signed-off-by: Linus Torvalds <torvalds@osdl.org>