Evaluate expressions fully when doing type comparisons on case statements
When we check the case statement type against the type of the switch
statement it is associated with, we need to make sure that the case
expression has been fully type-evaluated.

It should have happened automatically as we evaluate the switch
statement, but if that has not happened, we'd follow a NULL pointer due
to the unevaluated type.

Thanks to Adrian Bunk for reporting a segfault for reiser4

Signed-off-by: Linus Torvalds <torvalds@osdl.org>