When taking the address of a symbol, fix up the pointer type
The "volatile" and "const"'ness of the symbol will become part of the
pointer type, but we don't want to keep around things like "pointer to
unsigned", so strip that part off.

This still keeps other attributes, like "pointer to 'static'". Useful?
Probably not. I'll have to think about it.

Signed-off-by: Linus Torvalds <torvalds@osdl.org>