example: OP_COPY must destroy any old pseudo state
Toto, I have a feeling We're not in SSA land any more.

(Not that it really helps.  The example compiler not only had SSA
assumptions, it also depended on liveness information and insn->def
state, both of which are destroyed by the un-ssa phase.  So this is
really a small bandage around a much larger problem, and doesn't really
fix anything fundamental).

Signed-off-by: Linus Torvalds <torvalds@osdl.org>