prepare for #strong_{define,undef}
This patch renames symbol->weak to symbol->attr and adds
symbolic names for it's values.

I don't like the new name, and will be happy to rename if
you suggest me something different.

Probably it is also better to move ->attr into the NS_MACRO
part of the union and make it integer.

No changes in pre-process.o

Signed-off-by: Oleg Nesterov <oleg@tv-sign.ru>