[PATCH] beginning of SYM_RESTRICT rewrite: restricted_binop_type()
* new helper - restricted_binop_type(); gives resulting type of binary
operation when at least one of arguments is SYM_RESTRICT.  NULL is
returned in case of error.  Arguments will make sense when we get
to consolidation of compatible_binop_...() (next patch), switching
to better recovery from type errors for these guys (next after that)
and smarter handling of ~ on small restricted types.

* compatible_restricted_binop() and check_case_type() switched to
using that puppy